Fixes styleguide forms and several styleguide regressions
