+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### fixed
+- styleguide forms page breaks
+- several styleguide css regressions
 
 ## [0.8.3] 2017-06-12
 ### changed
