+++ components/forms/index.js
@@ -2,6 +2,7 @@ export { default as Form } from './Form';
 export { default as FormGroup } from './FormGroup';
 export { default as ModalFormGroup } from './ModalFormGroup';
 export { default as FormGroupError } from './FormGroupError';
+export { default as FormSummary } from './FormSummary';
 export { default as SubmitButton } from './SubmitButton';
 export { default as Input } from './Input';
 export { default as Checkbox } from './Checkbox';
