+++ components/navigation/Header.js
@@ -2,8 +2,10 @@ import React from 'react';
 
 
 export default function Header(props) {
+  const { className } = props;
+
   return (
-    <div className="Header">
+    <div className={`Header ${className}`}>
       <div className="MiniHeader">
         <div className="container">
           {props.miniHeader}
