+++ components/navigation/VerticalNav.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+export default function VerticalNav(props) {
+  const { children } = props;
+
+  return (
+    <div className="VerticalNav">
+      {children}
+    </div>
+  );
+};
