+++ scss/manager.scss
@@ -1,4 +1,6 @@
+// TODO: these need to be imported through shared package
 @import '../components/scss/linode-components';
+@import '../styleguide/scss/index';
 
 @import 'components/index';
 @import 'layout/index';
