+++ src/api/errors.js
@@ -1,0 +1,21 @@
+export async function reduceErrors(response) {
+  const json = await response.json();
+  const errors = { _: [] };
+  json.errors.forEach(error => {
+    let key = '_';
+    if (error.field) {
+      key = error.field;
+    }
+
+    if (error.field_crumbs) {
+      key += `.${error.field_crumbs}`;
+    }
+
+    const list = errors[key] || [];
+    list.push(error);
+    if (!errors[key]) {
+      errors[key] = list;
+    }
+  });
+  return errors;
+}
