--- src/components/forms/index.js
@@ -1,1 +1,0 @@
-export { reduceErrors, dispatchOrStoreErrors, FormSummary } from './FormSummary';
