+++ src/linodes/linode/backups/components/ScheduleForm.js
@@ -3,7 +3,8 @@ import React, { Component, PropTypes } from 'react';
 import { Form, FormGroup, FormGroupError, Select, SubmitButton } from 'linode-components/forms';
 
 import { linodes } from '~/api';
-import { dispatchOrStoreErrors, FormSummary } from '~/components/forms';
+import { dispatchOrStoreErrors } from '~/api/util';
+import { FormSummary } from 'linode-components/forms';
 
 
 export default class ScheduleForm extends Component {
