+++ src/profile/components/ChangeEmail.js
@@ -4,7 +4,8 @@ import { Card, CardHeader } from 'linode-components/cards';
 import { Form, FormGroup, FormGroupError, SubmitButton, Input } from 'linode-components/forms';
 
 import { profile } from '~/api';
-import { dispatchOrStoreErrors, FormSummary } from '~/components/forms';
+import { dispatchOrStoreErrors } from '~/api/util';
+import { FormSummary } from 'linode-components/forms';
 
 
 export default class ChangeEmail extends Component {
