+++ styleguide/package.json
@@ -7,6 +7,5 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
-  "license": "ISC"
+  "author": ""
 }
