+++ styleguide/scss/navigation.scss
@@ -1,10 +1,8 @@
 .Styleguide {
-  &-nav {
-    padding: 10px 4px 0 4px;
-  }
 
   &-nav ul {
     list-style-type: none;
+    padding: 0;
   }
 
   &-nav li {
